- renaming shapes.rs to shape.rs  - moving shapes.rs in extra folder
